Split the advisor feature into a separate page
